refactor(extension): extract Go binary resolution to dedicated module
